requirements/requirements-normal.txt
ipykernel==6.29.5
pandas==2.2.2
Pillow==10.3.0
tqdm==4.66.4
opencv-python==4.11.0
albumentations==2.0.6